Re-added manual add TTSlot functionality
